+++ pkg/gen/input/precommit/internal/file/pre_commit_action.go
@@ -1,0 +1,34 @@
+package file
+
+import (
+	_ "embed"
+	"path/filepath"
+
+	"github.com/giantswarm/devctl/v7/pkg/gen/input"
+	"github.com/giantswarm/devctl/v7/pkg/gen/input/precommit/internal/params"
+)
+
+//go:generate go run ../../../update-template-sha.go pre-commit-action.yaml.template
+//go:embed pre-commit-action.yaml.template
+var createPreCommitActionTemplate string
+
+//go:embed pre-commit-action.yaml.template.sha
+var createPreCommitActionTemplateSha string
+
+func NewCreatePreCommitActionInput(p params.Params) input.Input {
+	return input.Input{
+		Path:         filepath.Join(p.Dir, ".github", "workflows", "zz_generated.pre-commit.yaml"),
+		TemplateBody: createPreCommitActionTemplate,
+		// Use non-default delimiters so Go's template engine does not interpret
+		// the GitHub Actions ${{ }} expressions in the file content.
+		TemplateDelims: input.InputTemplateDelims{Left: "[[", Right: "]]"},
+		TemplateData: map[string]interface{}{
+			"Header":       params.Header("#", createPreCommitActionTemplateSha),
+			"Language":     p.Language,
+			"HasBash":      params.HasFlavor(p, "bash"),
+			"HasMd":        params.HasFlavor(p, "md"),
+			"HasHelmchart": params.HasFlavor(p, "helmchart"),
+			"RepoName":     p.RepoName,
+		},
+	}
+}